+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Binary package now built with cpack.
 * Added Conan option `with_save` to enable/disable 
   load/save support. 
+* Added support for ArduinoJson for baremetal.
 
 1.6.2 [24/07/24]
 * Deprecated config options `ramOffset`, `ramSize`,